VIC-35: Make all data accessible per tenant changes after review
