+++ src/main/java/de/caritas/cob/agencyservice/api/repository/TenantAware.java
@@ -1,0 +1,13 @@
+package de.caritas.cob.agencyservice.api.repository;
+
+/**
+ * Marker interface for entities that need to support tenant feature.
+ */
+public interface TenantAware {
+
+  /**
+   * Sets the fields tenantId during save operations.
+   * @param tenantId
+   */
+  void setTenantId(Long tenantId);
+}
