--- src/main/java/de/caritas/cob/agencyservice/api/repository/TenantSupport.java
@@ -1,6 +1,0 @@
-package de.caritas.cob.agencyservice.api.repository;
-
-public interface TenantSupport {
-
-  void setTenantId(Long tenantId);
-}
