+++ src/main/java/de/caritas/cob/agencyservice/api/repository/diocese/Diocese.java
@@ -1,6 +1,6 @@
 package de.caritas.cob.agencyservice.api.repository.diocese;
 
-import de.caritas.cob.agencyservice.api.repository.TenantSupport;
+import de.caritas.cob.agencyservice.api.repository.TenantAware;
 import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -29,7 +29,7 @@ import org.hibernate.annotations.ParamDef;
 @Setter
 @FilterDef(name = "tenantFilter", parameters = {@ParamDef(name = "tenantId", type = "long")})
 @Filter(name = "tenantFilter", condition = "tenant_id = :tenantId")
-public class Diocese implements TenantSupport {
+public class Diocese implements TenantAware {
 
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE)
