+++ src/test/java/de/caritas/cob/agencyservice/api/service/AgencyServiceTenantAwareIT.java
@@ -1,11 +1,13 @@
-package de.caritas.cob.agencyservice.api.admin.service;
+package de.caritas.cob.agencyservice.api.service;
 
 import de.caritas.cob.agencyservice.AgencyServiceApplication;
+import de.caritas.cob.agencyservice.api.exception.MissingConsultingTypeException;
 import de.caritas.cob.agencyservice.api.tenant.TenantContext;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -15,7 +17,10 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = AgencyServiceApplication.class)
@@ -23,9 +28,12 @@ import org.springframework.transaction.annotation.Transactional;
 @AutoConfigureTestDatabase(replace = Replace.ANY)
 @DirtiesContext(classMode = ClassMode.BEFORE_CLASS)
 @TestPropertySource(properties = "multitenancy.enabled=true")
-@Transactional
 @Sql(value = "/setTenants.sql", executionPhase = ExecutionPhase.BEFORE_TEST_METHOD)
-public class AgencyAdminServiceTenantSupportIT extends AgencyAdminServiceIT {
+@Transactional(propagation = Propagation.NEVER)
+public class AgencyServiceTenantAwareIT extends AgencyServiceITBase {
+
+  @Autowired
+  private PlatformTransactionManager transactionManager;
 
   @Before
   public void beforeEach() {
@@ -38,35 +46,19 @@ public class AgencyAdminServiceTenantSupportIT extends AgencyAdminServiceIT {
   }
 
   @Test
-  public void saveAgency_Should_PersistsAgency() {
-    super.saveAgency_Should_PersistsAgency();
+  public void getAgencies_Should_returnMatchingAgencies_When_postcodeAndConsultingTypeIsGiven()
+      throws MissingConsultingTypeException {
+    super.getAgencies_Should_returnMatchingAgencies_When_postcodeAndConsultingTypeIsGiven();
   }
 
   @Test
-  public void saveAgency_Should_SetOfflineToTrue_WhenPersistsAgency() {
-    super.saveAgency_Should_SetOfflineToTrue_WhenPersistsAgency();
+  public void setAgencyOffline_Should_FlagAgencyAsOfflineAndSetUpdateDate() {
+    super.setAgencyOffline_Should_FlagAgencyAsOfflineAndSetUpdateDate();
   }
 
   @Test
-  public void saveAgency_Should_ProvideValidAgencyLinks() {
-    super.saveAgency_Should_ProvideValidAgencyLinks();
-  }
-
-  @Test
-  public void updateAgency_Should_PersistsAgencyChanges() {
-    super.updateAgency_Should_PersistsAgencyChanges();
-  }
-
-  @Test
-  public void updateAgency_Should_ProvideValidAgencyLinks() {
-    super.updateAgency_Should_ProvideValidAgencyLinks();
-  }
-
-  @Test
-  public void getAgency_Should_returnExpectedAgency_When_agencyWithIdExists() {
-    super.getAgency_Should_returnExpectedAgency_When_agencyWithIdExists();
+  public void getAgenciesByConsultingType_Should_returnResults_When_ConsultingTypeIsValid() {
+    super.getAgenciesByConsultingType_Should_returnResults_When_ConsultingTypeIsValid();
   }
 
 }
-
-
