+++ src/test/resources/database/AgencyDatabase.sql
@@ -48,7 +48,7 @@ CREATE SEQUENCE SEQUENCE_AGENCY_POSTCODE_RANGE
     START WITH 100000
     INCREMENT BY 1;
 
-INSERT INTO DIOCESE (ID, TENANT_ID, NAME, ID_OLD, CREATE_DATE, UPDATE_DATE) VALUES (0, null, 'Aachen', 1, '2018-12-11 11:18:20', '2019-04-08 11:28:57');
+INSERT INTO DIOCESE (ID, TENANT_ID, NAME, ID_OLD, CREATE_DATE, UPDATE_DATE) VALUES (0, 1, 'Aachen', 1, '2018-12-11 11:18:20', '2019-04-08 11:28:57');
 INSERT INTO DIOCESE (ID, TENANT_ID, NAME, ID_OLD, CREATE_DATE, UPDATE_DATE) VALUES (1, null, 'Köln', 2, '2018-12-11 11:18:20', '2018-12-11 11:18:20');
 INSERT INTO DIOCESE (ID, TENANT_ID, NAME, ID_OLD, CREATE_DATE, UPDATE_DATE) VALUES (2, null, 'Münster', 3, '2018-12-11 11:18:20', '2018-12-11 11:18:20');
 INSERT INTO DIOCESE (ID, TENANT_ID, NAME, ID_OLD, CREATE_DATE, UPDATE_DATE) VALUES (3, null, 'Bamberg', 4, '2018-12-11 11:18:20', '2018-12-11 11:18:20');
